Front-end test call to trail weather, reformatted views
